+++ UltiTools-API/src/main/java/com/ultikits/ultitools/UltiTools.java
@@ -279,22 +279,4 @@ public final class UltiTools extends JavaPlugin implements Localized {
         }
         return config;
     }
-
-    /**
-     * get UltiTools UUID
-     *
-     * @return UUID
-     * @throws IOException if an I/O error occurs
-     */
-    public static String getUltiToolsUUID() throws IOException {
-        File dataFile = new File(UltiTools.getInstance().getDataFolder(), "data.json");
-        JSON json = new cn.hutool.json.JSONObject();
-        if (dataFile.exists()) {
-            json = JSONUtil.readJSON(dataFile, StandardCharsets.UTF_8);
-        } else {
-            json.putByPath("uuid", IdUtil.simpleUUID());
-            json.write(new FileWriter(dataFile));
-        }
-        return json.getByPath("uuid").toString();
-    }
 }
